@@ -8,22 +8,19 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.1.3",
-        "bensampo/laravel-enum": "^1.17",
-        "fideloper/proxy": "^4.0",
-        "intervention/image": "^2.4",
-        "jeroennoten/laravel-adminlte": "^1.25",
-        "laravel/framework": "5.7.*",
-        "laravel/tinker": "^1.0",
-        "spatie/laravel-permission": "^2.37"
+        "php": "^7.2.5",
+        "fideloper/proxy": "^4.2",
+        "fruitcake/laravel-cors": "^1.0",
+        "guzzlehttp/guzzle": "^6.3",
+        "laravel/framework": "^7.0",
+        "laravel/tinker": "^2.0"
     },
     "require-dev": {
-        "beyondcode/laravel-dump-server": "^1.0",
-        "filp/whoops": "^2.0",
-        "fzaninotto/faker": "^1.4",
-        "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^2.0",
-        "phpunit/phpunit": "^7.0"
+        "facade/ignition": "^2.0",
+        "fzaninotto/faker": "^1.9.1",
+        "mockery/mockery": "^1.3.1",
+        "nunomaduro/collision": "^4.1",
+        "phpunit/phpunit": "^8.5"
     },
     "config": {
         "optimize-autoloader": true,
