@@ -1,5 +1,7 @@
 <?php
 
+use Illuminate\Support\Facades\Route;
+
 /*
 |--------------------------------------------------------------------------
 | Web Routes
@@ -14,8 +16,6 @@
 Route::get('/', function () {
     return view('welcome');
 });
-
-Auth::routes();
 
 Route::middleware('auth')->group(function() {
 
